+++ src/Client/ConfirmationStatus.php
@@ -1,0 +1,28 @@
+<?php
+
+namespace CrazyGoat\RabbitStream\Client;
+
+class ConfirmationStatus
+{
+    public function __construct(
+        private readonly bool $confirmed,
+        private readonly ?int $errorCode = null,
+        private readonly ?int $publishingId = null,
+    ) {
+    }
+
+    public function isConfirmed(): bool
+    {
+        return $this->confirmed;
+    }
+
+    public function getErrorCode(): ?int
+    {
+        return $this->errorCode;
+    }
+
+    public function getPublishingId(): ?int
+    {
+        return $this->publishingId;
+    }
+}