+++ src/Client/ProducerConfig.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace CrazyGoat\RabbitStream\Client;
+
+class ProducerConfig
+{
+    /** @var ?callable */
+    public readonly mixed $onConfirmation;
+
+    public function __construct(
+        public readonly ?string $name = null,
+        ?callable $onConfirmation = null,
+    ) {
+        $this->onConfirmation = $onConfirmation;
+    }
+}